+++ include/roq/client/event_log_reader.hpp
@@ -54,6 +54,10 @@ struct ROQ_PUBLIC EventLogReader {
 
     // position manager
     virtual void operator()(Event<PortfolioUpdate> const &) = 0;
+
+    // risk management
+    virtual void operator()(Event<RiskLimits> const &) = 0;
+    virtual void operator()(Event<RiskLimitsUpdate> const &) = 0;
   };
 
   virtual ~EventLogReader() {}